# Massive truncate() performance gain.