```
name ............................... min time      avg time    std dv   runs
truncate .......................... 11.765 ms     13.770 ms    ±1.686   x360
truncate2 .......................... 0.025 ms      0.041 ms    ±0.024  x1000
truncate3 .......................... 0.025 ms      0.033 ms    ±0.009  x1000
```

Signed-off-by: David Krause <enthus1ast@users.noreply.github.com>